Various
git-svn-id: http://dcplusplus.svn.sourceforge.net/svnroot/dcplusplus/dcplusplus/trunk@658 f586c4b0-9d0d-0410-bc51-f1fe239209fc

+++ client/NmdcHub.cpp
@@ -737,7 +737,7 @@ void NmdcHub::onLine(const string& aLine) throw() {
 string NmdcHub::checkNick(const string& aNick) {
 	string tmp = aNick;
 	for(size_t i = 0; i < aNick.size(); ++i) {
-		if(tmp[i] <= 32 || tmp[i] == '|' || tmp[i] == '$' || tmp[i] == '<' || tmp[i] == '>') {
+		if(static_cast<u_int8_t>(tmp[i]) <= 32 || tmp[i] == '|' || tmp[i] == '$' || tmp[i] == '<' || tmp[i] == '>') {
 			tmp[i] = '_';
 		}
 	}
